7 ai prompt and api design (#70)
* removed breakit.txt

* Success:

It's scraping! We're returning the data successfully and in the shape we want.
We should think about filtering by location
Gotta write a util method to add the data to the DB

* docstrings

* updated headless command. No more popups

* GH qa:

Greenhouse QA is returning data as expected

* add the ChatGPT library and framework

* add error handling

* create chatGPT file and make adjustments

* add adjust the function

* add filter logic

* utils/gpt:

added sample job description to utils.py
retooled prompt in gpt

* Utils/GPT:

Wrote SQL query to update the json_response with GPT data
Converted json_response column to JSONB for easier mutation

* prompt and select_us_roles:

Adjusted the prompt - further testing needed
Added exclusions for select_us_roles to remove countries outside US
Also excluding a couple job boards for testing purposes. those can be removed later

* prompt:

The longer, more detailed prompts seem to return the most consistant data (as far as GPT reading a job as entry level or nah)
Added a function to open a new tab for the job being scanned for testing

* GPT/Utils.py:

---------

Co-authored-by: Steven <stzheng716@gmail.com>