75 database (#81)
* touched some stuff up

* updated bulk insert job_boards & moved functionality out of the sqlite_db.py

* job_boards bulk insert

BS for JBs is functional. It successfully scans all the job boards and adds them to the database in one mass insert

* T2 bulk insert complete:

* removed SQLite files. Updated Migrate.py:

migrate.py drops all tables
migrate_JP.py drops and builds the job_postings table and keeps the job_boards table as is

* bulk insert set up for Job Descriptions. needs test

* bulk insert job_descriptions succesfully insert all jobs

* gpt insert:

Seems to be functioning well. Need to look at the output to make sure we're getting what we want

* gpt insert is absolutly solid